update float--->double in entry
对于一个entry，其中的基本类型为double，支持更大的数据范围，方便以后时间戳的加入。

+++ library/src/main/java/com/linheimx/app/library/adapter/IValueAdapter.java
@@ -5,5 +5,5 @@ package com.linheimx.app.library.adapter;
  */
 
 public interface IValueAdapter {
-    String value2String(float value);
+    String value2String(double value);
 }
